switch mysql to sql server

+++ Server/SunriseServer.Core/Models/Product.cs
@@ -6,7 +6,7 @@ namespace SunriseServerCore.Models
     {
         // basic info
         public int ProductID { get; set; }
-        public float Price { get; set; } = 0;
+        public double Price { get; set; } = 0;
         public string Image { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
